requirements.txt
python-dotenv
beautifultable
numpy
pandas
requests
